Datasets/evalplus/Java/src/test/java/com/example/HumanEval_61Test.java
package com.example;
import static org.junit.Assert.*;
import org.junit.Test;

public class HumanEval_61Test {

    @Test
    public void testCorrectBracketing1() {
        assertFalse(HumanEval_61.correctBracketing("("));
    }

    @Test
    public void testCorrectBracketing2() {
        assertTrue(HumanEval_61.correctBracketing("()"));
    }

    @Test
    public void testCorrectBracketing3() {
        assertTrue(HumanEval_61.correctBracketing("(()())"));
    }

    @Test
    public void testCorrectBracketing4() {
        assertFalse(HumanEval_61.correctBracketing(")(()"));
    }

    @Test
    public void testCorrectBracketing5() {
        assertTrue(HumanEval_61.correctBracketing("()()(()())()"));
    }

    @Test
    public void testCorrectBracketing6() {
        assertTrue(HumanEval_61.correctBracketing("()()((()()())())(()()(()))"));
    }

    @Test
    public void testCorrectBracketing7() {
        assertFalse(HumanEval_61.correctBracketing("((()())))"));
    }

    @Test
    public void testCorrectBracketing8() {
        assertFalse(HumanEval_61.correctBracketing("(((("));
    }

    @Test
    public void testCorrectBracketing9() {
        assertFalse(HumanEval_61.correctBracketing("(()"));
    }

    @Test
    public void testCorrectBracketing10() {
        assertFalse(HumanEval_61.correctBracketing("()()(()())())(()"));
    }

    @Test
    public void testCorrectBracketing11() {
        assertTrue(HumanEval_61.correctBracketing("()()()()()"));
    }

    @Test
    public void testCorrectBracketing12() {
        assertFalse(HumanEval_61.correctBracketing("((())))("));
    }

    @Test
    public void testCorrectBracketing13() {
        assertFalse(HumanEval_61.correctBracketing(")()()))("));
    }

    @Test
    public void testCorrectBracketing14() {
        assertTrue(HumanEval_61.correctBracketing("((((()))))"));
    }

    @Test
    public void testCorrectBracketing15() {
        assertFalse(HumanEval_61.correctBracketing(")))))))"));
    }

    @Test
    public void testCorrectBracketing16() {
        assertFalse(HumanEval_61.correctBracketing("()(())(()()()(()(()(()((((((((((())()())))))))(()))())((()(()(())())())())())())(())))())(()(()(())(()()(((())))())(()))()(())(())()(()())"));
    }

    @Test
    public void testCorrectBracketing17() {
        assertFalse(HumanEval_61.correctBracketing("((())((())())((((())))()()(()(()(()())())())((((())()((((((())()())))))(("));
    }

    @Test
    public void testCorrectBracketing18() {
        assertTrue(HumanEval_61.correctBracketing("()(())((()()(((())))())(()))()(()())"));
    }

    @Test
    public void testCorrectBracketing19() {
        assertFalse(HumanEval_61.correctBracketing(")"));
    }

    @Test
    public void testCorrectBracketing20() {
        assertTrue(HumanEval_61.correctBracketing(""));
    }

    @Test
    public void testCorrectBracketing21() {
        assertFalse(HumanEval_61.correctBracketing("((())((())"));
    }

    @Test
    public void testCorrectBracketing22() {
        assertFalse(HumanEval_61.correctBracketing(")()())()()))()()()()()()()"));
    }
}

// End of Test
